Split chain manager module into multiple files.
